Rewrite authentication to use the state parameter.
We no longer have a single static auth url, instead we generate a new one for each connection.

The state parameter is now stored in a cookie and used to verify that it's the same client.

Also the PKCE challenge is now per-client, and associated with the state parameter.

Also added a bunch of comments and reorganization to make life easier.